feat: add windows support
